requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.9.2
pydantic-settings==2.5.2
python-multipart==0.0.12
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-dotenv==1.0.1
httpx==0.27.2
pytest==8.3.3
pytest-asyncio==0.24.0
websockets==13.1
supabase==2.15.1
asyncpg==0.30.0
sqlalchemy[asyncio]==2.0.40
google-genai>=0.1.0
langgraph
langchain-core
